# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Auto Anime Downloader daemon",
        "title": "Auto Anime Downloader API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8091",
    "basePath": "/api/v1",
    "paths": {
...
        "/standalone-animes": {
            "post": {
                "description": "Adds a standalone anime and immediately downloads the episodes already aired",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "standalone"
                ],
                "summary": "Track an anime that is not in any AniList list",
                "parameters": [
                    {
                        "description": "Media ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.StandaloneAnimeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.StandaloneAnimeAddResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/standalone-animes/{id}": {
            "delete": {
                "description": "Removes the standalone record; optionally deletes the downloaded episodes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "standalone"
                ],
                "summary": "Stop tracking a standalone anime",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AniList media ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Delete the downloaded episodes (default false)",
                        "name": "delete_episodes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.StandaloneAnimeAddResponse": {
            "type": "object",
            "properties": {
                "added": {
                    "description": "Added e quantos episodios entraram na fila agora. 0 e resposta normal: um anime\nNOT_YET_RELEASED (ou um cujo proximo episodio ainda nao foi ao ar) fica acompanhado do\nmesmo jeito. Quem traduz isso e o toast, nao um codigo de erro.",
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "api.StandaloneAnimeRequest": {
            "type": "object",
            "properties": {
                "media_id": {
                    "type": "integer",
                    "example": 21
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/api.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}